requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.10.0
psycopg2>=2.9.3,<2.10
cryptography==42.0.5
factory_boy==3.2.0
pillow==10.4.0
drf-yasg>=1.21.4
djangorestframework_simplejwt==5.5.0